refactor: move vote recap in separate files
