+++ SmoothPoll/views.py
@@ -1,13 +1,13 @@
 from django.http import Http404
 from django.shortcuts import render
-from django.urls import reverse
 from apps.polls_management.models.poll_model import PollModel
-import configparser
 
 from apps.polls_management.views.poll_short_id_view import PollShortIdView
 from apps.votes_results.classes.vote_consistency.check_consistency_session import CheckConsistencySession
-from apps.votes_results.views.majority_judgment_vote_view import SESSION_CONSISTENCY_CHECK
-from apps.votes_results.views.single_option_vote_view import SESSION_SINGLE_OPTION_VOTE_ID
+
+# todo: remove this dependency 
+from apps.votes_results.views.vote.majority_judgment_vote_view import SESSION_CONSISTENCY_CHECK
+from apps.votes_results.views.vote.single_option_vote_view import SESSION_SINGLE_OPTION_VOTE_ID
 
 def home(request):
     """
