+++ apps/votes_results/views/__init__.py
@@ -1,6 +1,8 @@
-from .single_option_vote_view import *
-from .majority_judgment_vote_view import *
+from .vote.single_option_vote_view import *
+from .vote.majority_judgment_vote_view import *
 from .votable_polls_view import *
 from .generic_vote_view import *
 from .results.single_option_results_view import *
 from .results.majority_judgment_results_view import *
+from .vote_recap.single_option_recap_view import *
+from .vote_recap.majority_judgment_recap_view import *
